demo/spaces/requirements.txt
gradio>=4.44.0
torch>=2.3.0
transformers>=4.46.0
peft>=0.12.0
accelerate>=0.34.0
pillow>=10.0.0
qwen-vl-utils>=0.0.8
